use uint32_t for column indexes

+++ metagraph/src/annotation/annotation_converters.hpp
@@ -60,7 +60,7 @@ convert_to_greedy_BRWT(RowDiffColumnAnnotator &&annotation,
 template <class StaticAnnotation>
 std::unique_ptr<StaticAnnotation>
 convert_to_BRWT(const std::vector<std::string> &annotation_files,
-                const std::vector<std::vector<uint64_t>> &linkage_matrix,
+                const std::vector<std::vector<uint32_t>> &linkage_matrix,
                 size_t num_parallel_nodes = 1,
                 size_t num_threads = 1,
                 const std::filesystem::path &tmp_dir = "");
